refactor: 수익률 관련 클래스로 분리
수익률 관련된 부분을 클래스로 분리
