+++ src/App.js
@@ -1,7 +1,8 @@
 const MissionUtils = require("@woowacourse/mission-utils");
 const Lotto = require("./Lotto");
-const MatchingNumber = require("./MatchingNumber");
 const WinnerNumber = require("./WinnerNumber");
+const MatchingNumber = require("./MatchingNumber");
+const ProfitRate = require("./ProfitRate.");
 
 class App {
   constructor() {
@@ -71,28 +72,9 @@ class App {
   }
 
   getLottoProfitRate(numberOfMathcingNumbers) {
-    const totalPrizeMoney = this.calculateTotalPrizeMoney(numberOfMathcingNumbers);
-    const lottoProfitRate = this.calculateProfitRate(totalPrizeMoney, this.payMoney);
+    const lottoProfitRate = new ProfitRate(numberOfMathcingNumbers, this.payMoney).getProfitRate();
     MissionUtils.Console.print(`총 수익률은 ${lottoProfitRate}%입니다.`);
     MissionUtils.Console.close();
-  }
-
-  calculateProfitRate(totalPrizeMoney, payMoney) {
-    return ((Math.round((totalPrizeMoney / payMoney) * 1000) / 1000) * 100).toFixed(1);
-  }
-
-  calculateTotalPrizeMoney(numberOfMathcingNumbers) {
-    return this.calculatePrizeMoney(numberOfMathcingNumbers).reduce((totalMoney, prizeMoney) => {
-      return (totalMoney += prizeMoney);
-    }, 0);
-  }
-
-  calculatePrizeMoney(numberOfMathcingNumbers) {
-    const currency = [5000, 50000, 1500000, 30000000, 2000000000];
-    const prizeMoney = numberOfMathcingNumbers.map((lotto, index) => {
-      if (lotto !== 0) return lotto * currency[index];
-    });
-    return prizeMoney.filter((money) => money !== undefined);
   }
 }
